feat: add a tryToActivateAgain entity list
This list contains entities that are active but for some reason (out of
DMA time?) could not be activated.
The engine will try to activate the entity again in the next display
frame.

Had to change the frameSetId of the example entities in the test to 0,
as not every entity would load because of the DMA time limit.