design: change Entity Velocity data type to 1:15:16fp
The original velocity data type (1:7:16fp) was chosen so I could
optimize velocity comparisons by just comparing the upper two bytes and
ignoring the loss of precision.

I changed my mind after writing out the initial design of the physics
subsystem. It turns out that such velocity comparisons are rarely
preformed when compared to sign checking, addition and subtraction,
which are easier to code using a 32 but data type. Testing with
the Horizontal Shooter demo has shown no noticeable speed changes.

Therefore I have changed the velocity to 1:15:16. I am aware that the
MSB of the velocity is always going to be `0x00` or `0xff`, but I
believe that the advantage in ease of coding is defiantly worth it.