feat: show memory usage table if SHOW_MEMORY_USAGE is defined
Prints a table showing the starting address, size, used memory, and free
space for each Data and RAM block of memory while assembling if
SHOW_MEMORY_USAGE is defined.

This functionality requires the developer to call the `finalizeMemory()`
macro at the end of the program or else the usage table will be invalid.

The unit test Makefile has also been updated to include a `memory-usage`
target to quickly view the memory usage table.