fix: Find and rewrite constants that have unknown values
The new version of my bass fork is stricter on evaluations than the
previous one and detected some constants with unknown values.

The old version of bass-untech treated unknown constants as `pc()`,
which can generate an invalid binary.

The code has been updated so that it assembles without error.

Analysis shows the following occurred because of the bug:

 * `MetaSprite.Vram._UploadDynamicTileset` was called on *every*
   MetaSprite activation. This went unnoticed because
   `_UploadDynamicTileset` has a test to prevent duplicate uploads.

 * Entity.lists.N_LISTS was set to -2175908 during the query phase of
   the assembler. This caused the bass-untech to not build any entity
   test data during the query phase.

   This bug went unnoticed because the entity test data was processed
   before entity unit tests. Had it been the other way around,
   bass-untech would not have been able to assemble the unit tests.