refactor: clean up scoping inside macros
Now that macros scope by default there is no need to put a namespace or
function inside a macro.

Macros that create functions (ie, the unit testing macros) have been
changed to inline.

Macros that have to open a namespace have been simplified so that the
scoped name is smaller.