feat: add static tiles to MetaSprites with dynamic tileset lifespan
The static tiles are uploaded to the end of the MetaSprite's VRAM
slot(s) on activation. Common tiles that are used in the MetaSprite's
frame tileset can now be moved into the static tileset in order to
reduce the number of tiles uploaded during VBlank when the MetaSprite's
frame changes.