fix: use correct values when decrementing Dma.transfersLeft
Use an expression to calculate the correct value that Dma.transfersLeft
should be decreased when queuing a Dma transfer to VBlank.

The CPU cycles used by each of the upload routines have been counted to
remove all guesswork from the calculations.

The Dma time used by the Dma.Tile16 module was found to be incorrect and
could have possibly lead to DMA overrun. It has been replaced with a
lookup table.