fix: remove lag when redrawing screen in MetaTile render test
A missing `WaitFrame` call between redrawing the screen and the select
test created an unintended loop where the map was constantly redrawn
until an NMI interrupt occurred between the `EnableDisplay_Full` call
and `JOY1` read.

Moving the `WaitFrame` call to the top of the loop causes `JOY1` to be
updated and the system to stop redrawing the map when the select button
is released.