Rewrite H-Trapezium to draw buffer in two DrawLine calls
When the number of lines to draw is > 127 then the buffer will be
shifted over by one byte with `mvp` in order to insert the second count
byte.

This will allow further optimisations to be applied to the DrawLine
routine as it no longer needs to worry about resuming the line.