Fix unsigned overflow in CalcLineXpos
This commit rearranges CalcLineXpos to minimize the risk of an unsigned
overflow within CalcLineXpos's multiplication.

It is still possible to CalcLineXpos to overflow, but I doubt it will be
a problem in practice.

This bug could be triggered by creating a TriangularTrapeziumWindow with
the coordinates (0x20, 0xa0), (0x08, 0x08), (0xf0, 0xd0).