Refactor Entity game loop
After reading through the Superfeather SNES Game Engine source code[1],
I realised that the old entity loop was inefficient and was unnecessary
complicated.  This lead to a new design of the entity loop that is
heavily inspired by super feather's object loop.

This commit changes the gameloop to a spaghetti coded loop. Where the
entity Process routine does not invoke the RTS instruction, but instead
jumps to a `Entity.GotoNextEntity` label.  This label will continue the
game loop, advancing to the next entity and calling the Process routine
if it exists.

The Entity GameLoop is no longer responsible for deactivating or
deleting entities that fall outside the active window.  Instead the
entity developer is expected to jump to one of the following code
blocks:

 * `Entity.DeactivateIfOutsideAndGotoNextEntity`
 * `Entity.DeactivateAndGotoNextEntity`
 * `Entity.DeleteIfOutsideAndGotoNextEntity`
 * `Entity.DeleteAndGotoNextEntity`

These code blocks will preform the desired tests, delete or deactivate
the entity if necessary and then jump to `Entity.GotoNextEntity`.

[1]: https://gitlab.com/cosmicchipsocket/superfeather