Refactor to include listId in BaseEntityStruct
I have decided that it would be nice to be able to call methods on
deactivated entities (for example, in game scripts).  This requires the
entity `BaseEntityStruct.functionTable` variable to always point to a
valid function table.

This necessitates the removal of the `_ft_currentList` hack and the
creation of a new entity variable, `listId`.

The `listId` variable is set to `BaseEntityData.initialListId` before
the entity's `Init` routine is called.  The `Init` method is allowed to
edit this value and when `Init` returns the spawning code will place the
entity in the entity list specified by `listId`.

As entity lists are now parsed by id number instead of a raw pointer
address, the `Entity.ChangeEntityListYAndGotoNextEntity` routine has
been changed to `Entity.ChangeEntityListIdAndGotoNextEntity`, which
accepts an entity list id as its argument.

To ensure correctness this new code raises a INVALID_ENTITY_LIST_ID
break whenever it encounters an invalid `listId` value.