Entity Init determines if an entity is always activated at spawn
The commit changes the `Entity.Spawn` routine so that the Entity's
`Init` routine is responsible for determining if an entity will always
activate (even if it is outside the active window).

The unit test entities have been updated so every entity will always be
activated at spawn, except for DeactivateIfOutsideEntity which will only
activated if it is inside the active window.

The Entity Spawn_Offscreen unit test has been rewritten to reflect the
new behaviour of Entity.Spawn.