Refactor to remove Entity Render Loop
This commit removes the Entity `Render` routine and makes the Entity
`Process` routine responsible for drawing the MetaSprites.

To help facilitate this change, a series of `@EndEntityProcess` routines
have been written that will process the MetaSprite Animation and draws
the entity's MetaSprite frame.

The entities used by the interactive tests have been updated to use the
new @EndEntityProcess routines.

None of the test entities invoke the new `DrawAndGotoNextEntity` code as
they were not used by the MetaSprite unit tests.