Fix location of Dma.ObjPalette.dirty flag
The previous location of `Dma.ObjPalette.dirty` flag was in wram7e.

When the NMI handler reads the flag using Data Bank 0 and thus the wrong
address is read.

The reason the unit tests passed is because the incorrect byte read was
non-zero and the ObjPalette was always transferred to CGRAM (even if I
didn't want it to).