Move entity to end of free list when it is deleted
A new variable `lastFreeEntity` has been added to the entity subsystem
to track the last entity in the free list.

To ensure this value is always correct a new test has been added to
`Entity.ValidateEntityLists`.

Two more `_TestEntityListsValid` calls have been added to the unit tests
in places where the `lastFreeEntity` was modified but not previously
checked.