Remove CountEnemiesWithId and CountActiveEnemies
Both of these function require the use of hard-coded entity list names.
I have decided that the are unnecessary with the current entity rewrite.

`CountEnemiesWithId` has been removed because it is slow, I doubt I will
need it and it's name will be confusing when the entityId table has been
implemented.

`Entity.CountActiveEnemies` has been removed because it can be easily
replaced with the `Entity.CountEntitiesInList` macro.