Add entityId to the BaseEntity struct
Each entity spawned by `Entity.Spawn` will be assigned a single byte
entityId value.

Using `Entity.GetEntityWithIdOrBranchIfEmpty`, an entityId can be
converted to an entity address, if it still exists.  If the entity
no-longer exists then the macro will branch given label, forcing the
programmer to handle the case where the entity is deleted.

The entityIdTable contains a setting `minimumEntityId` which contains
the minimum value an entityId.  This will allow a future expansion to
allocate a block of entityIds and ensure they will never be allocated by
`Entity.Spawn`.