Dupe GetEntityWithIdOrBranchIfEmpty for X or Y output
I initially had GetEntityWithIdOrBranchIfEmpty return the address to the
Y register, as that made it constant with the other entity routines.
However I just discovered that the `inc` and `dec` instructions only
work on the X index register.

This commit duplicates the GetEntityWithId… macros so the developer can
decide which register will contain the entity address.