Refactor Entity.Spawn entityId into entityDataId
This change will help separate the names of the three different id
variables used by the entity system.

To make the names consistent throughout the engine the EN_EntityList ROM
block has been renamed to EN_EntityDataList.

The names EntityDataId and EntityDataList were chosen after a few days
of brainstorming as it was the best I could come up with.