Revert "Add routine to delete entities outside rect/map"
This reverts commit b9553d2d0121ed372ef375a353de5cca622c3e3d.

I was planning on how to handle the level transitions when I realised
that keeping entities alive between levels would cause a wide variety of
issues regarding entityIds that are used or referenced by the level
scripts and/or entity structs.

To prevent these issues the new design is going to reset and
reinitialize the Entity and MetaSprite subsystems whenever a level is
loaded.

As the `DeleteEntitiesOutsideMetaTileMap` routine will no longer be
needed the commit that introduces it has been reverted.