Refactor to move VBlank code into WaitFrame routine
It came to my attention that the NMI handler did not handle lag frames
properly.  Under the old system a lag frame can dmaTime to be reset in
the middle of a processing frame, which can result in graphical glitches
and a VBlank overrun.

Looking though my code I realised I only needed the VBlank code to
execute in the middle of a `WaitFrame` call so I refactored the system
to do just that.

Design considerations:

 * frameCounter is now four bytes in size and will include any lag
   frames.
 * The NMI handler just increments a counter.
 * `WaitFrame` VBlank is split into three segments:
    * FinalizeFrame
    * VBlank
    * ResetQueues
 * If COUNT_FREE_CYCLES is defined, then WaitFrame will count the number
   of free cycles before VBlank start and save the result in
   `freeCycles`.
 * `ForceBlank` invokes WaitFrame to flush the buffer and prevent screen
   tearing.
 * `ForceBlank` disables any IRQ interrupts that may be active.
 * `Dma.Tile16.VBlank` is only executed if there is something in the
   buffer.