Add END_SLOPE tile collision
The END_SLOPE tile is placed at the end of each slope.  There is no
collision in the X-Axis but has SOLID collision for the bottom-center
and top-center tiles in the Y-Axis.

This new collision type was created to prevent a slope clip on a 45°
slope where xVecl > 3px/frame and the bottom/top is on the thinnest part
of the slope.