Change MetaTile map grid to column-major order with a fixed height
This commit changes the map grid format to a column-major array with a
fixed map height.  This removes the need for a multiplication when
preforming a position to tile conversion.

Column-major grid order was chosen because column based tile tests are a
lot more common then row based ones.

Testing with bsnes-plus shows that this change speeds up the
`MoveEntityAndProcessTileCollisions` by 9% when falling and 13% when
standing (compared to the previous commit).