Rewrite AnimatedTileset to use a shared decompression buffer
This commit rewrites the Resources.AnimatedTileset subsystem so it
decompresses the animated tileset data into the next free location
within Resources.decompressionBuffer.

One advantage of the new design is the elimination on the data size of
the animated tiles (aka MAX_ANIMATED_TILES_BLOCK_SIZE).

This commit is written in preparation for adding the
Resources.BackgroundImages subsystem to the engine.