Fix loading incorrect scene layout
This commit fixes a bug where `Resources.LoadScene` would use an
incorrect scene layout.

The bug was caused by me forgetting to mask the accumulator in the
`_SetSceneLayout` macro.

Previous testing did not catch this bug as the scenes had either a
palette id of 0 (which masked the input to `_SetSceneLayout`) or a
layout id of 0 (and thus were loaded normally).

To prevent this bug from happening again `_SetSceneLayout`the system
will break if it encounters an invalid scene layout id