Move functionTable from BaseEntityRomStruct into word address array
This commit removes functionTable from BaseEntityRomStruct and replaces
it with a word address array (Project.EntityFunctionTables).

This allows the editor to compile the EntityRomData into a byte vector.

For `Entity.Spawn` and `Entity._SpawnRoomEntity` this change does not
change execution time.

For entities spawned by action points this code will be a bit slower,
but the slowdown will be eliminated in the future when the projectiles
subsystem is written.