Change MetaSprite PaletteList to indexes within PaletteData
This commit changes the MetaSprite data format so the PaletteList
contains indexes within PaletteData.  Using indexes instead of addresses
reduces the number of pointers that need to be emitted by
untech-compiler.

Interestingly, this commit has a slightly faster
`MetaSprite.Palette._ProcessSlots` routine as a 16 bit lda/sta chain is
faster than the `mvn` instruction.

This commit also right aligns the SpikedWalker sprite palettes with a
garbage colour to test that all 15 colours are copied into CGRAM.