Fix broken Dma Tile16 unit tests
untech-editor moved `Project.DMA_Tile16Data` from the `RES_Lists` block
to a `RES_Block*`.  This broke the Tile16 tests as the `Dma.Tile16` code
can no longer access the `DmaTest.Tile16` test data.

This commit fixes the broken tests by creating a duplicate Tile16 module
that references the unit test data in its TileListData struct.

This fix is an ugly hack, but it works with minimal changes to the
engine.