Fix Tile16 queue not accessing tiles at end of a 4MiB ROM
As the MSB of the bank byte in the Tile16 queue is always clear, the
queue would not be able to access tiles stored in the last two banks in
a 4MiB ROM (banks 0xfe & 0xff).

This commit sets the MSB of the bank byte so that the queue always
points to the upper half of the address space (banks 0x80-0xff).

This commit also updates the Tile16 unit tests to verify the queue's
bank and tileHalfSize bytes are valid.