Add 10 unit test entity lists (and remove the Platforms entity list)
This commit adds 10 new entity lists (EntityList_0 to EntityList_9) to
the unit testing project.

The non interactive tests have been updated to use the new entity lists.
All references to the Player, Entity, or Projectile entity lists have
been removed from the non-interactive tests.

The number 10 was chosen because it is a nice round number and because
it increases the number of entity lists tested by the
`GameLoop_ChangeList` test.

This commit was made as preparation for the removal of the Player and
Projectile entity lists from the engine.

This commit also removes the Platforms entity list as it is no longer
used by the tests.