Optimize 16 and 32 bit Division routines
While watching Ben Eater's "Binary to decimal can’t be that hard,
right?" video[1], I decided to compare his division algorithm with the
one I was using and noticed it was better.

The new algorithm uses less ROM space on 16 bit division and it's faster
when 5 or more bits of the result are set.

For 32 bit division, the new algorithm uses 3 more bytes of ROM space,
but it's faster when 3 or more bits of the result are set.

[1]: https://www.youtube.com/watch?v=v3-a-zqKfgA