Fix tile collision clip when moving upwards
This commit fixes a bug where an upwards moving entity can clip into a
SOLID tile above it.  When this happens, the entity would be ejected out
of the SOLID tile in both the X and Y axes on the next frame.

This bug was introduced to the engine in commit
0023325f0d34d1294d4df6ca04ac041fadb2ef15