Rewrite crumbling tiles to use a circular buffer
Last night, after committing the crumbling tile code, I had a brainwave.
I could easily extend a circular buffer into a multi staged queue.

Profiling (with Mesen-S) shows this code is slightly faster than the old
code in the worse case (drawing 3 tiles from the last queue) but it is a
bit slower when both queues are empty.

                        | min       | max       | max       |
                        | Process   | Process   | Activate  |
    linked-list queue   |       914 |      8534 |       732 |
    circular buffer     |      1144 |      8516 |       628 |

I have decided to keep the circular buffer code, as it is a lot easier
to understand and has reduced memory requirements over the linked-list
queue.