Remove screen flashes in the Load_Room script test
The interactive Load_Room script test contained 3 annoying full-screen
black-white flashes in under a second.

This commit eliminates the flashes by disabling the screen for the
entire test.

The test has been moved to the non-interactive
`tests/scripting/scripting.inc` file as it no longer displays any
graphics to the screen.

This commit also speeds up the test by decreasing the distance the
test-player has to travel to touch the room script-trigger tiles.