Contribute soundtrack (and various other sound assets) for Rescue Game
I recorded it internally as SNESDev 2025 Space Station Game, as it was not yet
named. The project itself currently is named Rescue.

The SFX that was created has not actually been added to the collection yet to
assign it a slot and priority setting, as the game is still under development.