Fix broken tile collisions on the entity's first Process frame
`BaseEntity.MetaSprite.collisionBoxes` is set to 0 when an enemy is
spawned but that value cannot be null.  The code originally assumed this
value is set before any collision testing takes place.

This value is not set in the entity's `Init` function if it calls
`MetaSprite.Animation.SetAnimationId` as the sprite tiles have not been
uploaded to VRAM yet (as the entity might be offscreen).

The normal order of operations in an entity's `Process` function is:
 * Process tile collisions
 * Update state/animation if there is a tile collision
 * Process animation, which can set the MetaSprite frame and also sets
   the `collisionBoxes` variable
 * Draw entity

This results in the first `Process` frame using invalid collision box
data for the tile collisions.

This commit fixes the bug by populating the first `Data.CollisionBoxData`
instance with a blank collision box data so calling
`MoveEntityAndProcessTileCollisions` with `collisionBoxes = 0` does not
result in an invalid tile collision.

I did not notice this bug earlier because the enemies spawn offscreen or
when the screen is black.